requirements.txt
Django==3.2.14
djangorestframework==3.13.1
django-localflavor==3.1
phonenumbers==8.12.52